+++ src/builtin/widgets/mod.rs
@@ -2,6 +2,7 @@
 //!
 //! If the `widgets` feature is enabled, then built-in widgets will be automatically loaded and ready to use.
 
+pub mod checkbox;
 pub mod radio_button;
 pub mod scroll;
 pub mod slider;